+++ aviasales2/src/main/java/com/example/aviasales2/entity/Wallet.java
@@ -7,6 +7,7 @@ import lombok.Builder;
 import lombok.Data;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Data
@@ -18,18 +19,18 @@ public class Wallet {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
-    private Double sum;
+    private BigDecimal sum;
     @OneToOne(mappedBy = "wallet")
     private User owner;
     public Wallet(){
 
     }
 
-    public Double getSum() {
+    public BigDecimal getSum() {
         return sum;
     }
 
-    public void setSum(Double sum) {
+    public void setSum(BigDecimal sum) {
         this.sum = sum;
     }
